Testing routines optimized and stability increased

+++ scripts/Windows10_Privacy.ps1
@@ -1,6 +1,6 @@
 # Windows 10 privacy settings
 # Michi Dönselmann
-# Last change: 11.06.2020
+# Last change: 11.09.2021
 
 # Execution examples:
 # Enable privacy protection: powershell.exe -ExecutionPolicy Bypass "& '.\Windows10_Privacy.ps1 ' -enable:$true"
@@ -24,24 +24,12 @@ $tasks =  "Microsoft Compatibility Appraiser","ProgramDataUpdater","Consolidator
 
 # global error handling
 # only show defined error messages
-$ErrorActionPreference = "SilentlyContinue";
-$error.Clear();
-
-function errorhandling($message = $error[0].Exception.Message, $Code)
-{
-	# write execution Log to %temp%
-	$error | out-file -PSPath $env:temp\PowerShell_Execution_Error.log -Append;
-	# on error just show content of Exeption.Message
-	$message = "Error executing section $($code): " + $($message)
-	write-warning $($message)
-	# stop script on error
-	read-host
-	exit $code;
-}
+$ErrorActionPreference = "SilentlyContinue"
+$error.Clear()
 
 #----------------------------------------------------------------------------------
 # get system architecture if it's needed for further actions
-$arch = (Get-Process -Id $PID).StartInfo.EnvironmentVariables["PROCESSOR_ARCHITECTURE"];
+$arch = $env:PROCESSOR_ARCHITECTURE
 if ($arch -eq "AMD64")
 {
 	$arch = "x64"
@@ -53,7 +41,7 @@ else
 
 #----------------------------------------------------------------------------------
 # get windows edition
-$edition = get-windowsedition -online -ErrorAction SilentlyContinue
+$edition = get-windowsedition -online -ErrorAction $ErrorActionPreference
 
 #----------------------------------------------------------------------------------
 # check if script runs in admin context
@@ -63,7 +51,7 @@ if (-NOT ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdent
     Write-Warning "You have to run this Script with elevated privileges"
     Write-Host "Press any key to exit the script"
     read-host
-	exit 0;
+	exit 0
 }
 
 #--------------------------------
@@ -77,73 +65,65 @@ if ($enable -eq "True")
 
 	#----------------------------------------------------------------------------------
 	# disable data collecting services
-	foreach($a in $services)
+	foreach($service in $services)
 	{
-		set-service $a -startuptype disabled
-
-		# errorhandling
-		if ($error.Count -gt 0)
+		if (Get-Service $service)
 		{
-			errorhandling -code 1;
-		}	
+			write-host "Disable service '$service'"
+			set-service $service -startuptype disabled
+		}
 	}
 
 	#----------------------------------------------------------------------------------
 	# disable telemetry (also possible with GPO)
 	$path = "HKLM:\Software\Policies\Microsoft\Windows\DataCollection"
-	New-Item -path $path -Force
+	if (!(test-path $path)) 
+	{ 
+		New-Item -path $path	
+	}
 	
+	write-host "Configure telemetry"
 	if ($edition.edition -eq "Enterprise")
-        {
-                Set-ItemProperty -path $path -name AllowTelemetry -value "0" -type "DWord"
-        }
-        else
-        {
-                Set-ItemProperty -path $path -name AllowTelemetry -value "1" -type "DWord"
-        }
-		
-	# errorhandling
-	if ($error.Count -gt 0)
 	{
-		errorhandling -code 2;
+		Set-ItemProperty -path $path -name AllowTelemetry -value "0" -type "DWord"
+	}
+	else
+	{
+		Set-ItemProperty -path $path -name AllowTelemetry -value "1" -type "DWord"
 	}
 
 	#----------------------------------------------------------------------------------
 	# disable update delivery optimization (also possible with GPO)
-	$path = "HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeliveryOptimization"
-	Set-ItemProperty -path $path -name DODownloadMode -value "0" -type "DWord"
-		
-	# errorhandling
-	if ($error.Count -gt 0)
+	$path = "HKLM:\Software\Policies\Microsoft\Windows\DeliveryOptimization"
+	if (test-path $path)
 	{
-		errorhandling -code 3;
+		write-host "Configure delivery optimization"
+		Set-ItemProperty -path $path -name DODownloadMode -value "0" -type "DWord"
 	}
 		
 	#----------------------------------------------------------------------------------
 	# disable sheduled tasks
-	foreach($a in $tasks)
+	write-host "Disable scheduled tasks"
+	foreach($task in $tasks)
 	{
-		$task = Get-ScheduledTask $a;
-		Disable-ScheduledTask $task;
-		
-		# errorhandling
-		if ($error.Count -gt 0)
-		{
-			errorhandling -code 4;
+		$taskdetail = Get-ScheduledTask $task -ErrorAction $ErrorActionPreference
+		if ($taskdetail)
+		{	
+			Disable-ScheduledTask $taskdetail
 		}
 	}
+	write-host " "
 
 	#----------------------------------------------------------------------------------
 	# disable OneDrive for file sync (also possible with GPO)
 	$path = "HKLM:\SOFTWARE\Policies\Microsoft\Windows\OneDrive"
-	New-Item -path $path -Force
-	Set-ItemProperty -path $path -name DisableFileSyncNGSC -value "1"
-		
-	# errorhandling
-	if ($error.Count -gt 0)
+	if (!(test-path $path))
 	{
-		errorhandling -code 5;
+		New-Item -path $path
 	}
+	Write-Host "Disable OneDrive for file sync"
+	Set-ItemProperty -path $path -name DisableFileSyncNGSC -value "1"
+
 }
 else
 
@@ -156,61 +136,51 @@ else
 {
 	#----------------------------------------------------------------------------------
 	# enable data collecting services
-	foreach($a in $services)
+	foreach($service in $services)
 	{
-		set-service $a -startuptype automatic
-
-		# errorhandling
-		if ($error.Count -gt 0)
+		if (Get-Service $service)
 		{
-			errorhandling -code 1;
-		}	
+			write-host "Enable service '$service'"
+			set-service $service -startuptype automatic
+		}
 	}
 
 	#----------------------------------------------------------------------------------
 	# enable telemetry (also possible with GPO)
 	$path = "HKLM:\Software\Policies\Microsoft\Windows\DataCollection"
-	Remove-ItemProperty -path $path -name AllowTelemetry
-			
-	# errorhandling
-	if ($error.Count -gt 0)
+	if (test-path $path)
 	{
-		errorhandling -code 2;
+		write-host "Restore telemetry"
+		Remove-ItemProperty -path $path -name AllowTelemetry
 	}
 	
 	#----------------------------------------------------------------------------------
 	# enable update delivery optimization (also possible with GPO)
 	$path = "HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeliveryOptimization"
-	Remove-ItemProperty -path $path -name DODownloadMode
-		
-	# errorhandling
-	if ($error.Count -gt 0)
+	if (test-path $path)
 	{
-		errorhandling -code 3;
+		write-host "Restore delivery optimization"
+		Remove-ItemProperty -path $path -name DODownloadMode
 	}
 
 	#----------------------------------------------------------------------------------
 	# enable sheduled tasks
-	foreach($a in $tasks)
+	write-host "Enable scheduled tasks"
+	foreach($task in $tasks)
 	{
-		$task = Get-ScheduledTask $a;
-		Enable-ScheduledTask $task;
-		
-		# errorhandling
-		if ($error.Count -gt 0)
-		{
-			errorhandling -code 4;
+		$taskdetail = Get-ScheduledTask $task -ErrorAction $ErrorActionPreference
+		if ($taskdetail)
+		{	
+			Enable-ScheduledTask $taskdetail
 		}
 	}
 	
 	#----------------------------------------------------------------------------------
 	# enable OneDrive for file sync (also possible with GPO)
 	$path = "HKLM:\SOFTWARE\Policies\Microsoft\Windows\OneDrive"
-	Remove-ItemProperty -path $path -name DisableFileSyncNGSC
-		
-	# errorhandling
-	if ($error.Count -gt 0)
+	if (test-path $path)
 	{
-		errorhandling -code 5;
+		Write-Host "Restore OneDrive for file sync"
+		Remove-ItemProperty -path $path -name DisableFileSyncNGSC
 	}
 }
